@@ -58,8 +58,6 @@
   "overrides": {
     "@testing-library/dom": "$@testing-library/dom",
     "puppeteer": "$puppeteer",
-    "react": "$react",
-    "react-dom": "$react-dom",
     "storybook-addon-next-router": {
       "next": "$next"
     },
@@ -69,14 +67,14 @@
   "dependencies": {
     "@bugsnag/js": "^8.2.0",
     "@bugsnag/plugin-react": "^8.2.0",
-    "@clerk/nextjs": "^6.12.1",
+    "@clerk/nextjs": "^6.12.4",
     "@google-cloud/storage": "^7.15.0",
     "@hookform/resolvers": "^3.9.1",
     "@hubspot/api-client": "^11.2.0",
     "@mdx-js/loader": "^3.1.0",
     "@mux/mux-node": "^9.0.1",
     "@mux/mux-player-react": "3.1.0",
-    "@oaknational/oak-components": "^1.92.0",
+    "@oaknational/oak-components": "^1.93.0",
     "@oaknational/oak-consent-client": "^2.1.1",
     "@oaknational/oak-curriculum-schema": "^1.52.2",
     "@oaknational/oak-pupil-client": "^2.14.0",
@@ -106,7 +104,7 @@
     "lodash": "^4.17.21",
     "mathjax": "^3.2.2",
     "nanoid": "^5.0.8",
-    "next": "^14.2.23",
+    "next": "^15.2.1",
     "next-sanity-image": "^6.1.1",
     "next-seo": "^6.6.0",
     "node-fetch": "^2.6.7",
@@ -115,9 +113,9 @@
     "posthog-js": "^1.181.0",
     "posthog-node": "^4.2.1",
     "querystring-es3": "^0.2.1",
-    "react": "^18.3.1",
-    "react-aria": "^3.35.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "react-aria": "^3.38.0",
+    "react-dom": "^19.0.0",
     "react-focus-on": "^3.9.4",
     "react-hook-form": "^7.53.1",
     "react-stately": "^3.33.0",
@@ -154,11 +152,11 @@
     "@graphql-codegen/typescript-operations": "^4.4.0",
     "@netlify/edge-functions": "^2.11.1",
     "@netlify/plugin-nextjs": "^5.8.1",
-    "@next/bundle-analyzer": "^14.2.20",
-    "@next/env": "^14.2.20",
+    "@next/bundle-analyzer": "^15.2.1",
+    "@next/env": "^15.2.1",
     "@next/eslint-plugin-next": "^15.2.1",
     "@percy/cli": "v1.30.2",
-    "@prettier/plugin-xml": "3.2.2",
+    "@prettier/plugin-xml": "3.4.1",
     "@sanity/client": "^6.24.1",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/exec": "^6.0.3",
@@ -225,15 +223,15 @@
     "microdiff": "^1.5.0",
     "micromatch": "^4.0.8",
     "next-router-mock": "^0.9.13",
-    "next-sitemap": "^3.1.55",
+    "next-sitemap": "^4.2.3",
     "node-mocks-http": "^1.16.1",
     "open": "^10.1.0",
     "open-cli": "^8.0.0",
     "pa11y-ci": "^3.1.0",
     "postcss": "^8.4.47",
     "postcss-scss": "^4.0.9",
-    "prettier": "3.3.3",
-    "prettier-plugin-embed": "^0.4.15",
+    "prettier": "3.5.3",
+    "prettier-plugin-embed": "^0.5.0",
     "properties-file": "^3.5.9",
     "puppeteer": "^21.11.0",
     "rimraf": "^6.0.1",
@@ -248,8 +246,8 @@
     "svg-sprite": "^2.0.2",
     "ts-node": "^10.9.2",
     "tsx": "^4.19.2",
-    "typescript": "^5.6.2",
-    "typescript-plugin-css-modules": "^5.0.2",
+    "typescript": "^5.8.2",
+    "typescript-plugin-css-modules": "^5.1.0",
     "webpack": "^5.97.1",
     "webpack-bugsnag-plugins": "^1.8.0",
     "whatwg-fetch": "^3.6.20",
